+++ SRC/material/Yield/CMakeLists.txt
@@ -6,6 +6,8 @@
 #==============================================================================
 add_library(OPS_Material_YieldSurface OBJECT)
 
+target_sources(OPS_Runtime PRIVATE dispatch.cpp)
+
 target_include_directories(OPS_Material_YieldSurface PUBLIC ${CMAKE_CURRENT_LIST_DIR})
 
 add_subdirectory(./evolution/)
